DEFI-2665: move batch to separate file
